requirements.txt
# Core Trading Libraries
ccxt==4.2.9
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
python-dotenv==1.0.0

# Data Visualization
matplotlib==3.8.2
plotly==5.18.0
mplfinance==0.12.10b0

# Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
websocket-client==1.7.0

# Database
sqlalchemy==2.0.23
pymongo==4.6.1

# Configuration
pyyaml==6.0.1

# Notifications
python-telegram-bot==20.7

# Utilities
requests==2.31.0
schedule==1.2.0
pytz==2023.3.post1
colorama==0.4.6

# MetaTrader5 for Forex (Windows only)
MetaTrader5==5.0.45

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
